requirements.txt
Django>=5.0,<6.0
 djangorestframework>=3.15,<4.0
 djangorestframework-simplejwt>=5.3,<6.0
 django-environ>=0.11,<1.0
 django-cors-headers>=4.3,<5.0
 psycopg2-binary>=2.9,<3.0
 requests>=2.31,<3.0
 drf-spectacular>=0.27,<0.28
 django-filter>=24.2,<25.0
 google-generativeai>=0.7,<1.0
 pyfcm>=1.5,<2.0
 django-environ==0.11.2